requirements.txt
rhasspy-asr-kaldi
rhasspy-silence
rhasspy-hermes
paho-mqtt
flake8
pylint
